requirements.txt
# LangChain core
langchain
langchain_community
langchain_core
langgraph
langchain-chroma
langchain-huggingface
langchain-openai
langsmith

# Vector storage and embeddings
sentence-transformers
chromadb

# Data processing
pydantic
requests
python-dotenv

# Data visualization and analysis (new)
pandas
numpy
matplotlib
plotly